+++ packages/link/CHANGELOG.md
@@ -1,4 +1,4 @@
-# Change Log
+# @alfalab/core-components-link
 
 ## 5.3.4